VendeurService validation des données de l'inscription

+++ src/main/java/com/source/epoissonnerie/repository/VendeurRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VendeurRepository extends JpaRepository<VendeurPoisson,Long> {
+public interface VendeurRepository extends JpaRepository<VendeurPoisson, Integer> {
 }
